Move pagination gen to `GenApiModel`
